🔧 add(custom): 에러 메세지 constants 파일로 이동
