+++ src/apis/auth/post.ts
@@ -1,12 +1,6 @@
 import { ErrorResponseType } from '../../types/apis';
 import { typedPost } from '..';
-
-export const LOGIN_ERROR_MESSAGE: Record<string, string> = {
-  incorrect_password: '이메일 또는 비밀번호를 다시 확인해 주세요.',
-  resource_not_found: '올바른 이메일 주소를 입력해 주세요.',
-  already_exist_email: '이미 등록된 이메일입니다.',
-  invalid_request: '올바른 이메일 주소를 입력해 주세요.',
-};
+import { LOGIN_ERROR_MESSAGE } from './constants';
 
 // 공통 error 리턴
 const returnError = (err: unknown) => {
